After further discussion with Tai, the following changes have been made: - CamelCase script files have been renamed to conform to PEP 8 - The "extras" directory has been renamed to "templates" to be more apparent - The "sample_data" directory within START_HERE has been renamed to "fastq_files", and within it "sample_data_key.txt" has been renamed to "data_key.txt"
