+++ START_HERE/fastq_files/Lib303_test.fastq
