+++ START_HERE/fastq_files/Lib304_test.fastq
