+++ START_HERE/fastq_files/Lib305_test.fastq
