+++ START_HERE/fastq_files/Lib309_test.fastq
