+++ START_HERE/fastq_files/Lib310_test.fastq
