+++ START_HERE/fastq_files/Lib311_test.fastq
