+++ START_HERE/fastq_files/Lib312_test.fastq
