+++ START_HERE/fastq_files/Lib313_test.fastq
