+++ START_HERE/fastq_files/Lib314_test.fastq
