+++ START_HERE/samples.csv
@@ -1,1 +1,1 @@
-﻿Input FastQ Files,Sample/Group Name,Replicate number./sample_data/Lib303_test.fastq,N2,1./sample_data/Lib304_test.fastq,N2,2./sample_data/Lib305_test.fastq,N2,3./sample_data/Lib309_test.fastq,mut-16(pk710),1./sample_data/Lib310_test.fastq,mut-16(pk710),2./sample_data/Lib311_test.fastq,mut-16(pk710),3./sample_data/Lib312_test.fastq,prg-1(n4357),1./sample_data/Lib313_test.fastq,prg-1(n4357),2./sample_data/Lib314_test.fastq,prg-1(n4357),3
+﻿Input FastQ Files,Sample/Group Name,Replicate number./fastq_files/Lib303_test.fastq,N2,1./fastq_files/Lib304_test.fastq,N2,2./fastq_files/Lib305_test.fastq,N2,3./fastq_files/Lib309_test.fastq,mut-16(pk710),1./fastq_files/Lib310_test.fastq,mut-16(pk710),2./fastq_files/Lib311_test.fastq,mut-16(pk710),3./fastq_files/Lib312_test.fastq,prg-1(n4357),1./fastq_files/Lib313_test.fastq,prg-1(n4357),2./fastq_files/Lib314_test.fastq,prg-1(n4357),3
