+++ tests/testdata/counter/sample_sets/Lib309_short_fastq.csv
@@ -1,2 +1,2 @@
 Input FASTQ Files,Sample/Group Name,Replicate number
-../../../../START_HERE/sample_data/Lib309_test.fastq,mut-16(pk710),1
+../../../../START_HERE/fastq_files/Lib309_test.fastq,mut-16(pk710),1
