--- tiny/extras/samples.csv
@@ -1,1 +1,0 @@
-﻿Input FastQ Files,Sample/Group Name,Replicate number../../START_HERE/sample_data/Lib303_test.fastq,N2,1../../START_HERE/sample_data/Lib304_test.fastq,N2,2../../START_HERE/sample_data/Lib305_test.fastq,N2,3../../START_HERE/sample_data/Lib309_test.fastq,mut-16(pk710),1../../START_HERE/sample_data/Lib310_test.fastq,mut-16(pk710),2../../START_HERE/sample_data/Lib311_test.fastq,mut-16(pk710),3../../START_HERE/sample_data/Lib312_test.fastq,prg-1(n4357),1../../START_HERE/sample_data/Lib313_test.fastq,prg-1(n4357),2../../START_HERE/sample_data/Lib314_test.fastq,prg-1(n4357),3
