+++ tiny/templates/samples.csv
@@ -1,0 +1,1 @@
+﻿Input FastQ Files,Sample/Group Name,Replicate number../../START_HERE/fastq_files/Lib303_test.fastq,N2,1../../START_HERE/fastq_files/Lib304_test.fastq,N2,2../../START_HERE/fastq_files/Lib305_test.fastq,N2,3../../START_HERE/fastq_files/Lib309_test.fastq,mut-16(pk710),1../../START_HERE/fastq_files/Lib310_test.fastq,mut-16(pk710),2../../START_HERE/fastq_files/Lib311_test.fastq,mut-16(pk710),3../../START_HERE/fastq_files/Lib312_test.fastq,prg-1(n4357),1../../START_HERE/fastq_files/Lib313_test.fastq,prg-1(n4357),2../../START_HERE/fastq_files/Lib314_test.fastq,prg-1(n4357),3
