+++ tiny/templates/tinyrna-light.mplstyle
